Mise a jour code + documentation

+++ controllers/book.js
@@ -1,55 +1,65 @@
-// le controleur stocke toute la logique metier
-
 const Book = require('../models/book');
 const fs = require('fs');
+const sanitize = require('mongo-sanitize');
 
-// création d'un livre
-exports.createBook = (req, res, next) => {
-    const bookObject = JSON.parse(req.body.book);
+/**
+ * Création d'un Livre --------------------------------------------------
+ * @param {object} req requete
+ * @param {object} res reponse
+ */
+exports.createBook = (req, res) => {
+    const bookObject = JSON.parse(sanitize(req.body.book));
+
+    // controle de la note si pas entre 1 et 5 envoie une liste vide
+    if (bookObject.ratings[0].grade < 1 || bookObject.ratings[0].grade > 5) {
+        bookObject.ratings = [];
+    }
     const book = new Book({
         ...bookObject,
         userId: req.auth.userId,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${
             req.file.filename
         }`,
-        
     });
+
     book.save()
         .then(() => {
             res.status(201).json({ message: 'Objet enregistré !' });
         })
         .catch((error) => {
-            res.status(400).json({ error});
+            res.status(400).json({ error });
         });
 };
 
-// modifie un livre
-exports.modifyBook = (req, res, next) => {
+/**
+ * Modification d'un livre --------------------------------------------------
+ * @param {object} req requete
+ * @param {object} res reponse
+ */
+exports.modifyBook = (req, res) => {
     const bookObject = req.file
         ? {
-              ...JSON.parse(req.body.book),
+              ...JSON.parse(sanitize(req.body.book)),
               imageUrl: `${req.protocol}://${req.get('host')}/images/${
                   req.file.filename
               }`,
           }
-        : { ...req.body };
-
+        : { ...sanitize(req.body) };
     Book.findOne({ _id: req.params.id })
         .then((book) => {
+            // verification de l'authentification
             if (book.userId != req.auth.userId) {
                 res.status(403).json({ message: 'Not authorized' });
             } else {
-                const filename = book.imageUrl.split('/images/')[1];
-                fs.unlink(`images/${filename}`, () => {
-                    Book.updateOne(
-                        { _id: req.params.id },
-                        { ...bookObject, _id: req.params.id }
-                    )
-                        .then(() =>
-                            res.status(200).json({ message: 'Objet modifié!' })
-                        )
-                        .catch((error) => res.status(500).json({ error }));
-                });
+                // verification de la presence d'un fichier dans la requete
+                if (req.file) {
+                    const filename = book.imageUrl.split('/images/')[1];
+                    fs.unlink(`images/${filename}`, () => {
+                        updateBook(res, bookObject, req.params.id);
+                    });
+                } else {
+                    updateBook(res, bookObject, req.params.id);
+                }
             }
         })
         .catch((error) => {
@@ -57,8 +67,24 @@ exports.modifyBook = (req, res, next) => {
         });
 };
 
-// efface un  livre
-exports.deleteBook = (req, res, next) => {
+/**
+ * Mise à jour du livre dans la base de donnees
+ * @param {object} res reponse
+ * @param {object} bookObject livre
+ * @param {string} bookId Id du livre
+ */
+function updateBook(res, bookObject, bookId) {
+    Book.updateOne({ _id: bookId }, { ...bookObject, _id: bookId })
+        .then(() => res.status(200).json({ message: 'Objet modifié!' }))
+        .catch((error) => res.status(500).json({ error }));
+}
+
+/**
+ * Suppression d'un livre --------------------------------------------------
+ * @param {object} req requete
+ * @param {object} res reponse
+ */
+exports.deleteBook = (req, res) => {
     Book.findOne({ _id: req.params.id })
         .then((book) => {
             if (book.userId != req.auth.userId) {
@@ -81,58 +107,39 @@ exports.deleteBook = (req, res, next) => {
         });
 };
 
-// trouve tous les livres
-exports.findAllBooks = (req, res, next) => {
+/**
+ * Renvoie la liste de tous les livres --------------------------------------------------
+ * @param {object} req requete
+ * @param {object} res reponse
+ */
+exports.findAllBooks = (req, res) => {
     Book.find() // renvoie un tableau contenant toutes la collection
         .then((books) => res.status(200).json(books))
         .catch((error) => res.status(400).json({ error }));
 };
 
-// trouve un livre
+/**
+ * Trouve un livre avec son id en parametre --------------------------------------------------
+ * @param {object} req requete
+ * @param {object} res reponse
+ * @param {Function} next
+ */
 exports.findOneBook = (req, res, next) => {
-    
     Book.findOne({ _id: req.params.id })
         .then((book) => {
             res.status(200).json(book);
-            next()
+            next();
         })
         .catch((error) => res.status(404).json({ error }));
 };
 
-// note un livre
-exports.ratingBook = (req, res, next) => {
-    const bookId = req.params.id;
-    const userId = req.body.userId;
-    const rating = req.body.rating;
-
-    Book.findOneAndUpdate(
-        { _id: bookId, 'ratings.userId': { $ne: userId } }, // Vérifie si userId n'est pas déjà présent dans ratings
-        { $addToSet: { ratings: { userId: userId, grade: rating } } }, // Utilise $addToSet au lieu de $push
-        { new: true }
-    )
-        .then((book) => {
-            if (!book) {
-                // Si le livre n'est pas trouvé ou si l'utilisateur a déjà noté le livre, retourner une erreur
-                return res.status(404).json({ error: 'Le livre n\'a pas été trouvé ou l\'utilisateur a déjà noté le livre.' });
-            }
-            const averageBook = calculateAverageRating(book.ratings)
-            book.averageRating = averageBook
-            book.save() // Enregistre les modifications dans la base de données
-                .then((updatedBook) => {
-                    res.status(200).json(updatedBook);
-                    
-                })
-                .catch((error) => res.status(500).json({ error }));
-            
-        })
-        .catch((error) => res.status(500).json({ error }));
-};
-
-
-// Renvoie les 3 livres les mieux notes
-exports.bestRatingBooks = (req, res, next) => {
+/**
+ * Renvoie les 3 livres les mieux notes --------------------------------------------------
+ * @param {object} res reponse
+ */
+exports.bestRatingBooks = (req, res) => {
     Book.find()
-        .sort({ 'averageRating': -1 }) // Trie les livres par note décroissante
+        .sort({ averageRating: -1 }) // Trie les livres par note décroissante
         .limit(3) // Limite les résultats aux 3 premiers livres
         .then((books) => {
             res.status(200).json(books);
@@ -140,15 +147,55 @@ exports.bestRatingBooks = (req, res, next) => {
         .catch((error) => res.status(500).json({ error }));
 };
 
-// calcul la moyenne des notes d'un livre
+/**
+ * Notation d'un livre --------------------------------------------------
+ * @param {object} req requete
+ * @param {object} res reponse
+ */
+exports.ratingBook = (req, res) => {
+    const bookId = req.params.id;
+    const userId = req.auth.userId;
+    let rating = req.body.rating;
+
+    // verification de la plage de notation entre 0 et 5
+    if (rating > 5) {
+        rating = 5;
+    }
+
+    Book.findOneAndUpdate(
+        { _id: bookId, 'ratings.userId': { $ne: userId } }, // Vérifie si userId n'est pas déjà présent dans ratings
+        { $addToSet: { ratings: { userId: userId, grade: rating } } }, // Ajoute la notation si elle n'existe pas deja
+        { new: true }
+    )
+        .then((book) => {
+            if (!book) {
+                return res.status(404).json({
+                    error: "Le livre n'a pas été trouvé ou l'utilisateur a déjà noté le livre.",
+                });
+            }
+            const averageBook = calculateAverageRating(book.ratings);
+            book.averageRating = averageBook;
+            book.save()
+                .then((updatedBook) => {
+                    res.status(200).json(updatedBook);
+                })
+                .catch((error) => res.status(500).json({ error }));
+        })
+        .catch((error) => res.status(500).json({ error }));
+};
+
+/**
+ * Calcul la note moyenne d'un livre
+ * @param {Array} ratings
+ * @returns {number} average ou 0 si average vide
+ */
 const calculateAverageRating = (ratings) => {
     if (ratings.length === 0) {
         return 0; // Retourne 0 si le tableau est vide pour éviter une division par zéro
     }
 
-    const sum = ratings.reduce((total, rating) => total + rating.grade, 0);
+    const sum = ratings.reduce((total, rating) => total + rating.grade, 0); // somme de toutes les notes du livre
     const average = sum / ratings.length;
 
     return Math.round(average);
 };
-
